+++ Sources/Code/Engine/Audio/src/AudioUtils.cpp
@@ -5,9 +5,9 @@
 namespace {
 
 const float EXP_VOLUME_DYNAMIC_RANGE = 40.f;
-const float EXP_VOLUME_POWER = powf(10.f, EXP_VOLUME_DYNAMIC_RANGE / 20.f);
+const float EXP_VOLUME_POWER = pow(10.f, EXP_VOLUME_DYNAMIC_RANGE / 20.f);
 const float EXP_VOLUME_A = 1.f / EXP_VOLUME_POWER;
-const float EXP_VOLUME_B = logf(EXP_VOLUME_POWER);
+const float EXP_VOLUME_B = log(EXP_VOLUME_POWER);
 
 } // namespace
 
@@ -17,7 +17,7 @@ float CovertToExpVolume(float value) {
     if(value <= 0.001) {
         return 0.f;
     }
-    return EXP_VOLUME_A * expf(value * EXP_VOLUME_B);
+    return EXP_VOLUME_A * exp(value * EXP_VOLUME_B);
 }
 
 } // namespace Audio
